requirements.txt
amqp==5.3.1
asgiref==3.8.1
bcrypt==4.2.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.0
Django==5.1.5
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-filter==24.3
django-timezone-field==7.1
djangorestframework==3.15.2
future==1.0.0
idna==3.10
ipaddress==1.0.23
kombu==5.4.2
Markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
netmiko==4.5.0
ntc_templates==7.6.0
packaging==24.2
paramiko==3.5.0
prompt_toolkit==3.0.48
pycparser==2.22
pydantic==1.10.21
Pygments==2.19.1
pykeadhcp==0.6.0
PyNaCl==1.5.0
pynetbox==7.4.1
pyserial==3.5
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scp==0.15.0
six==1.17.0
sqlparse==0.5.3
textfsm==1.1.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
